# openapi.yaml
openapi: 3.0.3
info:
  title: SwiftTrack Airport Express Ticketing API
  description: Production-grade REST API for SwiftTrack Rail & Airport Transfer Booking System.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development / Docker Environment
paths:
  /api/v1/auth/register:
    post:
      summary: Register new customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName: { type: string }
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Successful registration with JWT tokens
  /api/v1/auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: JWT Access and Refresh Tokens
...